Refactors src/package.rs into multiple modules.
In general, one module should do one thing. The `src/package.rs` module
was doing too many things, including:

- compressing and decompressing packages
- handling package stores
- handling package names
- handling package types

Instead, I have split these into small modules that are more obvious in
what they do and can be more easily tested. In doing so, I have added
some unit tests as well.
